+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Tinify API Change Log
 =====================
 
+Under development
+-----------------
+* Enh: Creates component for image resizing
+
 1.0.0 August 07, 2017
 ---------------------
 * Initial release